Update menu, fptai apikey
