+++ ProjectX/Views/RegisterWindow.xaml
@@ -1,12 +1,14 @@
 ﻿<Window x:Class="ProjectX.Views.RegisterWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        Title="Register" Height="369" Width="325"
+        Title="Register" Height="411" Width="325"
         WindowStartupLocation="CenterScreen"
         ResizeMode="NoResize"
         Background="#202025" Icon="../assets/AppIcon.ico">
     <Grid Margin="10">
         <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
             <RowDefinition Height="Auto" />
@@ -24,15 +26,19 @@
         <Label Content="Password:" Grid.Row="3" VerticalAlignment="Center" />
         <PasswordBox x:Name="PasswordBox" Grid.Row="4" Margin="0,5" VerticalAlignment="Center" Foreground="#d8d9de" />
 
-        <Label Content="API Key:" Grid.Row="5" VerticalAlignment="Center" />
-        <TextBox Name="ApiKeyTextBox" Grid.Row="6" Margin="0,5,0,0" VerticalAlignment="Top" Foreground="#d8d9de" />
+        <Label Content="ZaloAI API Key:" Grid.Row="5" VerticalAlignment="Center" />
+        <TextBox Name="ZaloAI" Grid.Row="6" Margin="0,5,0,0" VerticalAlignment="Top" Foreground="#d8d9de" />
 
-        <Button Content="Đăng ký" Grid.Row="6" Click="RegisterButton_Click" Margin="0,50,0,0" VerticalAlignment="Top"
+        <Label Content="FPTAI API Key:" Grid.Row="7" VerticalAlignment="Center" />
+        <TextBox Name="FPTAI" Grid.Row="8" Margin="0,5,0,0" VerticalAlignment="Top" Foreground="#d8d9de" />
+
+        <Button Content="Đăng ký" Grid.Row="8" Click="RegisterButton_Click" VerticalAlignment="Center"
                 HorizontalAlignment="Center" Width="280" Background="#202123" BorderBrush="#ea4335"
                 HorizontalContentAlignment="Center" />
-        <Button Content="Đăng nhập" Grid.Row="6" Click="BackToLoginButton_Click" Margin="0,90,0,0"
+        <Button Content="Đăng nhập" Grid.Row="8" Click="BackToLoginButton_Click" Margin="0,95,0,0"
                 VerticalAlignment="Top"
                 HorizontalAlignment="Center" Width="280" Background="#202123" BorderBrush="#4285f4"
                 HorizontalContentAlignment="Center" />
+
     </Grid>
 </Window>
