+++ ProjectX/Views/TTS.xaml
@@ -1,0 +1,67 @@
+﻿<Window x:Class="ProjectX.Views.TTS"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        mc:Ignorable="d"
+        Title="Text-to-speech" Height="377" Width="728" WindowStartupLocation="CenterScreen" Background="#202025" ResizeMode="NoResize" Visibility="Visible" WindowStyle="SingleBorderWindow" LostTouchCapture="Grid_MouseLeftButtonDown" Loaded="Window_Loaded" MouseLeftButtonDown="Window_MouseLeftButtonDown" MouseDown="Window_MouseDown_1"
+        Icon="../assets/AppIcon.ico"
+        >
+
+    <Grid x:Name="_back" MouseLeftButtonDown="_back_MouseLeftButtonDown" Height="337" VerticalAlignment="Top">
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="57*"/>
+            <ColumnDefinition Width="79*"/>
+            <ColumnDefinition Width="502*"/>
+            <ColumnDefinition Width="90*"/>
+        </Grid.ColumnDefinitions>
+        <Grid.RowDefinitions>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+        <Label Content="Giọng:" HorizontalAlignment="Center" Margin="0,108,0,0" VerticalAlignment="Top" Height="27" Foreground="White" FontFamily="Segoe UI Semibold" Width="44"/>
+        <ComboBox x:Name="_nguoidoc" HorizontalAlignment="Left" Margin="3,108,0,0" VerticalAlignment="Top" Width="145" Foreground="White" FontFamily="Segoe UI Semibold" Height="25" Grid.ColumnSpan="2" Grid.Column="1">
+            <ComboBoxItem Content="Nữ miền Bắc" HorizontalAlignment="Left" Width="118"/>
+            <ComboBoxItem Content="Nam miền Bắc" HorizontalAlignment="Left" Width="118"/>
+            <ComboBoxItem Content="Nữ miền Nam" IsSelected="True" HorizontalAlignment="Left" Width="118"/>
+            <ComboBoxItem Content="Nam miền Nam" IsSelected="True" HorizontalAlignment="Left" Width="118"/>
+        </ComboBox>
+        <!-- Setting -->
+        <Label Content="Tốc độ :" HorizontalAlignment="Left" Margin="5,140,0,0" VerticalAlignment="Top" Height="27" Foreground="White" FontFamily="Segoe UI Semibold" Width="51"/>
+        <ComboBox x:Name="_tocdo" HorizontalAlignment="Left" VerticalAlignment="Top" Width="145" Margin="3,139,0,0" Foreground="White" FontFamily="Segoe UI Semibold" Height="25" Grid.ColumnSpan="2" Grid.Column="1">
+            <ComboBoxItem Content="Bình thường(1.0)" IsSelected="True" HorizontalAlignment="Left" Width="118"/>
+            <ComboBoxItem Content="Rất chậm(0.8)" HorizontalAlignment="Left" Width="118"/>
+            <ComboBoxItem Content="Chậm(0.9)" HorizontalAlignment="Left" Width="118"/>
+            <ComboBoxItem Content="Nhanh(1.1)" HorizontalAlignment="Left" Width="118"/>
+            <ComboBoxItem Content="Rất nhanh(1.2)" HorizontalAlignment="Left" Width="118"/>
+        </ComboBox>
+        <!-- Button -->
+        <Button x:Name="_run" Content="Đọc" HorizontalAlignment="Left" Height="24" Margin="10,179,0,0" VerticalAlignment="Top" Width="199" Click="_run_Click" Background="#202123" Foreground="White" FontFamily="Segoe UI Semibold" FontSize="12" BorderBrush="#34a853" Grid.ColumnSpan="3"/>
+        <Button x:Name="_download" Content="Download" HorizontalAlignment="Left" Height="24" Margin="10,206,0,0" VerticalAlignment="Top" Width="199" FontSize="12" Click="_download_Click" Background="#202123" BorderBrush="#4285f4" Grid.ColumnSpan="3"/>
+        <Button x:Name="_stop" Content="Dừng" HorizontalAlignment="Left" Height="24" Margin="10,232,0,0" VerticalAlignment="Top" Width="199" Click="_stop_Click" Foreground="White" Background="#202123" FontSize="12" FontFamily="Segoe UI Semibold" BorderBrush="#ea4335" Grid.ColumnSpan="3" />
+        <Button x:Name="btnOpenFile" Content="Mở file" Click="BtnOpenFile_Click" Height="24" Margin="10,257,0,0" VerticalAlignment="Top" Width="199" FontFamily="Segoe UI Semibold" FontSize="12" Background="#202123" BorderBrush="#fbbc05" HorizontalAlignment="Left" Grid.ColumnSpan="3"/>
+        <!-- Textbox -->
+        <TextBox Margin="95,0,10,0" TextWrapping="Wrap" VerticalScrollBarVisibility="Visible" AcceptsReturn="True" x:Name="_text" Foreground="#FFFFFF" Text="Nhập văn bản ở đây." BorderBrush="#FFFAFAFA" Background="#41414b" Height="227" VerticalAlignment="Center" Grid.ColumnSpan="2" Grid.Column="2" TextChanged="_text_TextChanged_1">
+            <TextBox.Resources>
+                <Style TargetType="{x:Type Border}">
+                    <Setter Property="CornerRadius" Value="10"/>
+                </Style>
+            </TextBox.Resources>
+            <TextBox.CaretBrush>
+                <LinearGradientBrush EndPoint="0.5,1" StartPoint="0.5,0">
+                    <GradientStop Color="#FF64B3F4"/>
+                    <GradientStop Color="#FFC2E59C" Offset="1"/>
+                </LinearGradientBrush>
+            </TextBox.CaretBrush>
+        </TextBox>
+
+        <!-- Tien trinh -->
+        <Image Source="../assets/Logo.png" HorizontalAlignment="Left" Margin="158,10,0,0" VerticalAlignment="Top" Height="52" Width="140" Grid.Column="2"/>
+        <ProgressBar x:Name="_process" Maximum="100" Minimum="0" Height="26" Margin="428,296,10,0" VerticalAlignment="Top" Grid.ColumnSpan="2" Grid.Column="2" Background="#34a853" BorderBrush="#34a853"/>
+        <Label Content="Tiến trình :" Margin="356,298,0,0" VerticalAlignment="Top" Height="21" Foreground="White" FontFamily="Segoe UI Semibold" RenderTransformOrigin="0.46,1.469" HorizontalAlignment="Left" Width="67" Grid.Column="2"/>
+        <Label x:Name="_tientring" Content="Chưa khởi động." HorizontalAlignment="Left" Margin="27,292,0,0" VerticalAlignment="Top" Height="27" Foreground="White" FontFamily="Segoe UI Semibold" Width="276" Grid.ColumnSpan="2" Grid.Column="1"/>
+        <Label Content="Trạng thái :" HorizontalAlignment="Left" Margin="10,292,0,0" VerticalAlignment="Top" Height="27" Foreground="White" FontFamily="Segoe UI Semibold" Width="69" Grid.ColumnSpan="2"/>
+        <Label x:Name="_kytu" Content="Ký tự đã nhập: 0" Margin="228,296,0,0" VerticalAlignment="Top" Height="26" Foreground="White" Background="#FF444444" RenderTransformOrigin="0.796,0.519" HorizontalAlignment="Left" Width="118" Grid.Column="2"/>
+        <Label Content="ZaloAI API Key:" HorizontalAlignment="Left" Margin="6,54,0,0" VerticalAlignment="Top" Height="27" Foreground="White" FontFamily="Segoe UI Semibold" FontWeight="Bold" Width="91" Grid.ColumnSpan="2"/>
+        <TextBox x:Name="_apikey" HorizontalAlignment="Left" Height="23" Margin="10,81,0,0" TextWrapping="Wrap" Text="(EMPTY)" VerticalAlignment="Top" Width="199" Foreground="White" Grid.ColumnSpan="3"/>
+    </Grid>
+</Window>
