+++ ProjectX/bin/Debug/zalo_tts/output.txt
@@ -1,1 +1,1 @@
-{"error_code":0,"error_message":"Successful.","data":{"url":"https://chunk.lab.zalo.ai/f7ba292a64448d1ad455/f7ba292a64448d1ad455"}}
+{"error_code":0,"data":{"url":"https://chunk.lab.zalo.ai/0fe0843ec7502e0e7741/0fe0843ec7502e0e7741"},"error_message":"Successful."}
